+++ src/main/java/kr/co/preq/domain/auth/dto/TokenRequestDto.java
@@ -1,4 +1,4 @@
-package kr.co.preq.domain.board.auth.dto;
+package kr.co.preq.domain.auth.dto;
 
 import javax.validation.constraints.NotBlank;
