+++ src/main/java/kr/co/preq/domain/auth/service/AuthService.java
@@ -1,9 +1,9 @@
-package kr.co.preq.domain.board.auth.service;
+package kr.co.preq.domain.auth.service;
 
 import java.util.UUID;
 
-import kr.co.preq.domain.board.auth.dto.AuthRequestDto;
-import kr.co.preq.domain.board.auth.dto.AuthResponseDto;
+import kr.co.preq.domain.auth.dto.AuthRequestDto;
+import kr.co.preq.domain.auth.dto.AuthResponseDto;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.core.Authentication;
@@ -12,8 +12,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 
-import kr.co.preq.domain.board.auth.dto.LogoutRequestDto;
-import kr.co.preq.domain.board.auth.dto.TokenRequestDto;
+import kr.co.preq.domain.auth.dto.LogoutRequestDto;
+import kr.co.preq.domain.auth.dto.TokenRequestDto;
 import kr.co.preq.domain.member.dto.MemberRequestDto;
 import kr.co.preq.domain.member.entity.Member;
 import kr.co.preq.domain.member.repository.MemberRepository;
